requirements.txt
tensorflow
Pillow
tensorflow_hub
fastapi